Чистит запрос
